Fixing binding tests
The visibility tests were trying to assert that the element had it's inline style prop "display" set to "inline-block" when the element was visible. That's not how it works. Only when the element is hidden, is the inline style prop "display: none" applied. Otherwise, there is no "display" inline style prop declared.

Fixed the tests to reflect this.